group_chat_service/requirements.txt
Flask==2.2.3
requests==2.28.2
flask-cors==3.0.10
werkzeug==2.2.3 
mysql-connector-python

